tests_package/requirements.txt
pytest
typing
pytest-playwright